add class GameLogic with simple GameLogic
Still needs a more elegant solution for casting

+++ EncounterMe/EncounterMe/Classes/Location.cs
@@ -14,7 +14,6 @@ namespace EncounterMe
         public float Latitude { get; set; }
         public float Longtitude { get; set; }
 
-        //public Location(int id, String name, double northCoord, double eastCoord)
         public Location(int ID, String Name, double Latitude, double Longtitude)
         {
             this.ID = ID;
@@ -23,7 +22,7 @@ namespace EncounterMe
             this.Longtitude = (float) Longtitude;
         }
 
-        public float distanceToUser(double lat, double lon)
+        public float distanceToUser(float lat, float lon)
         {
             return (float) (circumference * Math.Acos(Math.Sin(this.Latitude * Math.PI / 180.00) * Math.Sin(lat * Math.PI / 180.00) + Math.Cos(this.Latitude * Math.PI / 180.00) * Math.Cos(lat * Math.PI / 180.00) * Math.Cos((this.Longtitude - lon) * Math.PI / 180)));
         }
